correctly detect utf-8 encoding of text
The code that was detecting the encoding of `.txt` files was trying to
decode as cp-1252 before decoding as utf-8.  utf-8 will always validate
as cp-1252, but not vice-versa. Ergo, utf-8 with multibyte characters is
displayed as gibberish.  Move cp-1252 down the list until after utf-8
and utf-16.  There's more overlap between utf-16 and cp-1252, so that
may need to be adjusted later.