Update default MFR loading spinner to follow OSF style
Users prefer the pulsing ball to the COS-logo spinner.  Please see
https://centerforopenscience.github.io/osf-style/style.html#loader

This doesn't really affect the iframe content served by MFR. Instead,
this changes happens in the mfr.js and mfr.css files loaded from MFR
by the site hosting the iframe.  The fact that mfr.js is included in
most iframe content templates is a development artifact and will be
eventually removed.