Remove Unoconv renderer
The unoconv renderer is obsolete because handling of the conversion if
the file cannot be rendererd using the given rederer occurs in the
render handler. This allows any filetype to be rendered by any renderer,
provided there is a converter that is capable of converting the filetype
to a filtype that the given renderer can render.