Clarify why wont work locally
Added a clarifiaction on why the renderer won't work locally. The
rendering service needs to access the file thatis to be rendered, and
the url waterbutler gives the renderer is going to be inaccessible to
the outside world, unless the developer configures waterbutler to use an
externally accessible hostname/ip.