Dont pass empty header
The OSF provider was passing a header with an empty value, this change
means the if the header is empty, then the header isn't sent at all.

This technically is a small optimization as well, it removes a few
instructions that would get called inside aiohttp. if this header was
here. (This is probably insignificant, however.)