Set the background of the zoom image to "white"
Both the zoom image and the normal one underneath it are visible if
it has a transparent backgournd. This bug only affects non-exported
types: .GIF and .ICO. For exported ones, the export type .JPEG does
not have transparency so the background is set to white by default.
This fix simply applies such background when the zoom image's style
class is generated on the page.