CSV sniffer sniffs the full first row for dialect detection
CSV sniffer used to sniff a fixed size of 2048 Bytes of any given
file. However, this led to wrong guessed delimiter if the first
row is larger than 2048 Bytes. The issue is fixed and accuracy is
also improved by allowing the sniffer to adaptively sniff the 1st
full row as long as it is within the max render file size.